pyproject.toml
[project]
name = "assertion-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "grpcio>=1.76.0",
    "dspy-ai",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["design*", "output*"]

